lxd: Moves instance load and instance validation functions into instance pkg
Updates usage.
