+++ lxd/container_put.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/lxc/lxd/lxd/db"
 	deviceConfig "github.com/lxc/lxd/lxd/device/config"
+	"github.com/lxc/lxd/lxd/instance"
 	"github.com/lxc/lxd/lxd/operations"
 	"github.com/lxc/lxd/lxd/response"
 	"github.com/lxc/lxd/lxd/state"
@@ -42,7 +43,7 @@ func containerPut(d *Daemon, r *http.Request) response.Response {
 		return resp
 	}
 
-	c, err := instanceLoadByProjectAndName(d.State(), project, name)
+	c, err := instance.LoadByProjectAndName(d.State(), project, name)
 	if err != nil {
 		return response.NotFound(err)
 	}
@@ -115,12 +116,12 @@ func instanceSnapRestore(s *state.State, project, name, snap string, stateful bo
 		snap = name + shared.SnapshotDelimiter + snap
 	}
 
-	inst, err := instanceLoadByProjectAndName(s, project, name)
+	inst, err := instance.LoadByProjectAndName(s, project, name)
 	if err != nil {
 		return err
 	}
 
-	source, err := instanceLoadByProjectAndName(s, project, snap)
+	source, err := instance.LoadByProjectAndName(s, project, snap)
 	if err != nil {
 		switch err {
 		case db.ErrNoSuchObject:
