+++ lxd/storage_migration.go
@@ -353,7 +353,7 @@ func rsyncMigrationSink(conn *websocket.Conn, op *operations.Operation, args Mig
 				}
 
 				// Try and a load instance
-				s, err := instanceLoadByProjectAndName(args.Instance.DaemonState(),
+				s, err := instance.LoadByProjectAndName(args.Instance.DaemonState(),
 					args.Instance.Project(), snapArgs.Name)
 				if err != nil {
 					// Create the snapshot since it doesn't seem to exist
@@ -430,7 +430,7 @@ func rsyncMigrationSink(conn *websocket.Conn, op *operations.Operation, args Mig
 					}
 				}
 
-				_, err = instanceLoadByProjectAndName(args.Instance.DaemonState(),
+				_, err = instance.LoadByProjectAndName(args.Instance.DaemonState(),
 					args.Instance.Project(), snapArgs.Name)
 				if err != nil {
 					_, err = instanceCreateAsSnapshot(args.Instance.DaemonState(), snapArgs, args.Instance, op)
